move tests from lib.rs to tests folder
